Uncommit changes for deploy
